fix(footerbar): fix wcag issue of footerbar  closes2407
