+++ packages/geoview-core/src/ui/tabs/tab-panel.tsx
@@ -1,7 +1,8 @@
 /* eslint-disable react/require-default-props */
 import { ReactNode } from 'react';
 import { Box } from '@/ui';
-import { FocusTrapElement } from '@/core/components/common/focus-trap-element';
+import { FocusTrapContainer } from '@/core/components/common';
+import { TypeContainerBox } from '@/core/types/global-types';
 
 type TypeChildren = ReactNode;
 
@@ -13,6 +14,7 @@ export interface TypeTabPanelProps {
   value: number;
   id: string;
   children?: TypeChildren;
+  containerType?: TypeContainerBox;
 }
 
 /**
@@ -22,11 +24,13 @@ export interface TypeTabPanelProps {
  * @returns {JSX.Element} returns the tab panel
  */
 export function TabPanel(props: TypeTabPanelProps): JSX.Element {
-  const { children, value, index, id, ...other } = props;
+  const { children, value, index, id, containerType, ...other } = props;
 
   return (
     <Box role="tabpanel" hidden={value !== index} id={id} aria-labelledby={`simple-tab-${index}`} {...other}>
-      <FocusTrapElement id={`panel-${index}`} content={children} />
+      <FocusTrapContainer id={`panel-${index}`} containerType={containerType}>
+        {children}
+      </FocusTrapContainer>
     </Box>
   );
 }
